apps/worker/requirements.txt
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
redis==5.0.8
rq==1.16.2
pgvector==0.2.5
boto3==1.34.156
openai==1.42.0
pdfplumber==0.11.0
pypdf==4.3.1
python-docx==1.1.2
fastapi==0.115.8
pydantic==2.8.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
uvicorn==0.30.6
